backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
email-validator==2.1.0.post1
python-slugify==8.0.1
uv==0.1.14
